Improve project details button inside project table
