+++ src/components/project-list/project-table.tsx
@@ -13,8 +13,9 @@ import {
   StarTotal
 } from 'components/core/project'
 import { TagLabelGroup } from 'components/tags/tag-label'
-import { DropdownMenu, Button } from 'components/core'
+import { Button } from 'components/core'
 import { fromNow } from 'helpers/from-now'
+import { ProjectDetailsButton } from './project-details-button'
 
 type Props = {
   projects: BestOfJS.Project[]
@@ -95,45 +96,6 @@ const ProjectTableRow = ({
   )
   const showDownloads = sortOption.id === 'monthly-downloads'
   const showStars = !showDelta && !showDownloads
-
-  const getBookmarkMenuItem = () => {
-    if (!isLoggedIn) {
-      return { label: 'Add bookmark', icon: <GoBookmark />, disabled: true }
-    }
-    if (project.isBookmark) {
-      return {
-        label: 'Remove bookmark',
-        icon: <GoBookmark />,
-        onClick: () => removeBookmark(project)
-      }
-    }
-    return {
-      label: 'Add bookmark',
-      icon: <GoBookmark />,
-      onClick: () => addBookmark(project)
-    }
-  }
-
-  const getHomepageMenuItem = () =>
-    project.url && {
-      icon: <GoHome />,
-      label: 'Go to homepage',
-      url: project.url,
-      onClick: () => ({})
-    }
-
-  const items = [
-    // { type: 'label', label: 'Links' },
-    {
-      icon: <GoMarkGithub />,
-      label: 'Go to GitHub repository',
-      url: project.repository,
-      onClick: () => ({})
-    },
-    getHomepageMenuItem(),
-    { type: 'divider' },
-    getBookmarkMenuItem()
-  ]
 
   const toggleBookmark = () => {
     project.isBookmark ? removeBookmark(project) : addBookmark(project)
@@ -250,7 +212,7 @@ const ProjectTableRow = ({
 
       {showActions && (
         <ActionCell>
-          <DropdownMenu items={items} alignment="right" />
+          <ProjectDetailsButton project={project} isLoggedIn={isLoggedIn} />
         </ActionCell>
       )}
     </Row>
